optimizing highlighter code
