+++ setup.py
@@ -10,11 +10,15 @@ with open('README.rst', encoding="utf8") as readme_file:
 with open('HISTORY.rst', encoding="utf8") as history_file:
     history = history_file.read()
 
-requirements = ['click>=7.0', 'pdfminer.six>=20200726', 'lxml', 'python-docx', 'folia', 'pandas', 'camelot-py>=0.10.1', 'opencv-python>=4.5.5.62', 'pdftopng>=0.2.3']
+requirements = ['click>=7.0', 'pdfminer.six>=20200726', 'lxml', 'python-docx', 'folia',
+                'pandas', 'camelot-py>=0.10.1', 'opencv-python>=4.5.5.62', 'pdftopng>=0.2.3', 'pdftopng == 0.2.3',
+                'fitz == 0.0.1.dev2']
 
-setup_requirements = ['click>=7.0', 'pdfminer.six>=20200726', 'lxml', 'python-docx', 'folia', 'pandas', 'camelot-py>=0.10.1', 'opencv-python>=4.5.5.62', 'pdftopng>=0.2.3']
+setup_requirements = ['click>=7.0', 'pdfminer.six>=20200726', 'lxml', 'python-docx',
+                      'folia', 'pandas', 'camelot-py>=0.10.1', 'opencv-python>=4.5.5.62', 'pdftopng>=0.2.3']
 
-test_requirements = ['click>=7.0', 'pdfminer.six>=20200726', 'lxml', 'python-docx', 'folia', 'pandas', 'stanza', 'spacy', 'deepdiff', 'camelot-py>=0.10.1', 'opencv-python>=4.5.5.62', 'pdftopng>=0.2.3']
+test_requirements = ['click>=7.0', 'pdfminer.six>=20200726', 'lxml', 'python-docx', 'folia', 'pandas',
+                     'stanza', 'spacy', 'deepdiff', 'camelot-py>=0.10.1', 'opencv-python>=4.5.5.62', 'pdftopng>=0.2.3']
 
 setup(
     author="De Nederlandsche Bank",
@@ -31,7 +35,7 @@ setup(
         'Programming Language :: Python :: 3.8',
     ],
     description="Python package to convert spaCy and Stanza documents to NLP Annotation Format (NAF)",
-    long_description_content_type = 'text/x-rst',
+    long_description_content_type='text/x-rst',
     entry_points={
         'console_scripts': [
             'nafigator=nafigator.cli:main',
